Enhance Docker image layers architecture
Due to Docker image's layer architecture it worth it to place all the
things that changes more constantly at the bottom of the file, this
speeds up the build process. Thus this commits put at the top of the
file the install process of dependencies and at the bottom the source
code itself.